docs(habitat): atomize civ7 remainder rules
