--- .habitat/civ7/mapgen/domains/foundation/_remainder/prohibit_foundation_contract_config_bags/rule.json
@@ -1,95 +1,0 @@
-{
-  "schemaVersion": 2,
-  "id": "prohibit_foundation_contract_config_bags",
-  "title": "Prohibit Foundation Contract Config Bags",
-  "placement": {
-    "niche": "civ7/mapgen/domains/foundation",
-    "blueprint": "_remainder",
-    "category": "contract"
-  },
-  "operation": {
-    "kind": "check"
-  },
-  "ownerProject": "mod-swooper-maps",
-  "lane": "enforced",
-  "forbids": "foundation contracts importing domain config bags or using FoundationConfigSchema",
-  "why": "Foundation contracts must expose decomposed typed inputs rather than root config-bag surfaces.",
-  "remediate": null,
-  "message": "Remove foundation config-bag references from contract surfaces.",
-  "pathCoverage": [
-    {
-      "kind": "exact-path",
-      "patterns": [
-        "mods/mod-swooper-maps/src/domain/foundation/ops/compute-crust-evolution/contract.ts",
-        "mods/mod-swooper-maps/src/domain/foundation/ops/compute-crust/contract.ts",
-        "mods/mod-swooper-maps/src/domain/foundation/ops/compute-era-plate-membership/contract.ts",
-        "mods/mod-swooper-maps/src/domain/foundation/ops/compute-era-tectonic-fields/contract.ts",
-        "mods/mod-swooper-maps/src/domain/foundation/ops/compute-hotspot-events/contract.ts",
-        "mods/mod-swooper-maps/src/domain/foundation/ops/compute-mantle-forcing/contract.ts",
-        "mods/mod-swooper-maps/src/domain/foundation/ops/compute-mantle-potential/contract.ts",
-        "mods/mod-swooper-maps/src/domain/foundation/ops/compute-mesh/contract.ts",
-        "mods/mod-swooper-maps/src/domain/foundation/ops/compute-plate-graph/contract.ts",
-        "mods/mod-swooper-maps/src/domain/foundation/ops/compute-plate-motion/contract.ts",
-        "mods/mod-swooper-maps/src/domain/foundation/ops/compute-plate-topology/contract.ts",
-        "mods/mod-swooper-maps/src/domain/foundation/ops/compute-plates-tensors/contract.ts",
-        "mods/mod-swooper-maps/src/domain/foundation/ops/compute-segment-events/contract.ts",
-        "mods/mod-swooper-maps/src/domain/foundation/ops/compute-tectonic-history-rollups/contract.ts",
-        "mods/mod-swooper-maps/src/domain/foundation/ops/compute-tectonic-provenance/contract.ts",
-        "mods/mod-swooper-maps/src/domain/foundation/ops/compute-tectonic-segments/contract.ts",
-        "mods/mod-swooper-maps/src/domain/foundation/ops/compute-tectonics-current/contract.ts",
-        "mods/mod-swooper-maps/src/domain/foundation/ops/compute-tracer-advection/contract.ts",
-        "mods/mod-swooper-maps/src/recipes/standard/stages/foundation-lithosphere/steps/crust.contract.ts",
-        "mods/mod-swooper-maps/src/recipes/standard/stages/foundation-lithosphere/steps/plateGraph.contract.ts",
-        "mods/mod-swooper-maps/src/recipes/standard/stages/foundation-mantle/steps/mantleForcing.contract.ts",
-        "mods/mod-swooper-maps/src/recipes/standard/stages/foundation-mantle/steps/mantlePotential.contract.ts",
-        "mods/mod-swooper-maps/src/recipes/standard/stages/foundation-mantle/steps/mesh.contract.ts",
-        "mods/mod-swooper-maps/src/recipes/standard/stages/foundation-orogeny/steps/crustEvolution.contract.ts",
-        "mods/mod-swooper-maps/src/recipes/standard/stages/foundation-projection/steps/plateTopology.contract.ts",
-        "mods/mod-swooper-maps/src/recipes/standard/stages/foundation-projection/steps/projection.contract.ts",
-        "mods/mod-swooper-maps/src/recipes/standard/stages/foundation-tectonics/steps/plateMotion.contract.ts",
-        "mods/mod-swooper-maps/src/recipes/standard/stages/foundation-tectonics/steps/tectonics.contract.ts"
-      ]
-    }
-  ],
-  "scanRoots": [
-    "mods/mod-swooper-maps/src/domain/foundation/ops/compute-crust-evolution/contract.ts",
-    "mods/mod-swooper-maps/src/domain/foundation/ops/compute-crust/contract.ts",
-    "mods/mod-swooper-maps/src/domain/foundation/ops/compute-era-plate-membership/contract.ts",
-    "mods/mod-swooper-maps/src/domain/foundation/ops/compute-era-tectonic-fields/contract.ts",
-    "mods/mod-swooper-maps/src/domain/foundation/ops/compute-hotspot-events/contract.ts",
-    "mods/mod-swooper-maps/src/domain/foundation/ops/compute-mantle-forcing/contract.ts",
-    "mods/mod-swooper-maps/src/domain/foundation/ops/compute-mantle-potential/contract.ts",
-    "mods/mod-swooper-maps/src/domain/foundation/ops/compute-mesh/contract.ts",
-    "mods/mod-swooper-maps/src/domain/foundation/ops/compute-plate-graph/contract.ts",
-    "mods/mod-swooper-maps/src/domain/foundation/ops/compute-plate-motion/contract.ts",
-    "mods/mod-swooper-maps/src/domain/foundation/ops/compute-plate-topology/contract.ts",
-    "mods/mod-swooper-maps/src/domain/foundation/ops/compute-plates-tensors/contract.ts",
-    "mods/mod-swooper-maps/src/domain/foundation/ops/compute-segment-events/contract.ts",
-    "mods/mod-swooper-maps/src/domain/foundation/ops/compute-tectonic-history-rollups/contract.ts",
-    "mods/mod-swooper-maps/src/domain/foundation/ops/compute-tectonic-provenance/contract.ts",
-    "mods/mod-swooper-maps/src/domain/foundation/ops/compute-tectonic-segments/contract.ts",
-    "mods/mod-swooper-maps/src/domain/foundation/ops/compute-tectonics-current/contract.ts",
-    "mods/mod-swooper-maps/src/domain/foundation/ops/compute-tracer-advection/contract.ts",
-    "mods/mod-swooper-maps/src/recipes/standard/stages/foundation-lithosphere/steps/crust.contract.ts",
-    "mods/mod-swooper-maps/src/recipes/standard/stages/foundation-lithosphere/steps/plateGraph.contract.ts",
-    "mods/mod-swooper-maps/src/recipes/standard/stages/foundation-mantle/steps/mantleForcing.contract.ts",
-    "mods/mod-swooper-maps/src/recipes/standard/stages/foundation-mantle/steps/mantlePotential.contract.ts",
-    "mods/mod-swooper-maps/src/recipes/standard/stages/foundation-mantle/steps/mesh.contract.ts",
-    "mods/mod-swooper-maps/src/recipes/standard/stages/foundation-orogeny/steps/crustEvolution.contract.ts",
-    "mods/mod-swooper-maps/src/recipes/standard/stages/foundation-projection/steps/plateTopology.contract.ts",
-    "mods/mod-swooper-maps/src/recipes/standard/stages/foundation-projection/steps/projection.contract.ts",
-    "mods/mod-swooper-maps/src/recipes/standard/stages/foundation-tectonics/steps/plateMotion.contract.ts",
-    "mods/mod-swooper-maps/src/recipes/standard/stages/foundation-tectonics/steps/tectonics.contract.ts"
-  ],
-  "hookCheck": true,
-  "supportFiles": {
-    "baseline": ".habitat/civ7/mapgen/domains/foundation/_remainder/prohibit_foundation_contract_config_bags/baseline.json"
-  },
-  "runner": {
-    "name": "grit",
-    "files": {
-      "pattern": ".habitat/civ7/mapgen/domains/foundation/_remainder/prohibit_foundation_contract_config_bags/pattern.md"
-    },
-    "patternName": "prohibit_foundation_contract_config_bags"
-  }
-}
