+++ .habitat/civ7/mapgen/domains/foundation/rules/prohibit_foundation_op_contract_config_bags/baseline.json
