+++ .habitat/civ7/mapgen/domains/foundation/rules/prohibit_foundation_op_contract_config_bags/pattern.md
@@ -1,0 +1,20 @@
+---
+level: error
+---
+# Prohibit Foundation Op Contract Config Bags
+
+Foundation operation contracts must not depend on root config-bag imports or schemas.
+
+```grit
+language js(typescript)
+
+or {
+  import_statement(source=$source) where {
+    $source <: r".*@mapgen/domain/config.*"
+  },
+  `FoundationConfigSchema`
+} where {
+  $filename <: r".*mods/mod-swooper-maps/src/domain/foundation/ops/[^/]+/contract\.ts$"
+}
+```
+
