+++ .habitat/civ7/mapgen/domains/foundation/rules/prohibit_foundation_op_contract_config_bags/rule.json
@@ -1,0 +1,41 @@
+{
+  "schemaVersion": 2,
+  "id": "prohibit_foundation_op_contract_config_bags",
+  "title": "Prohibit Foundation Op Contract Config Bags",
+  "placement": {
+    "niche": "civ7/mapgen/domains/foundation",
+    "blueprint": "_self",
+    "category": "contract"
+  },
+  "operation": {
+    "kind": "check"
+  },
+  "ownerProject": "mod-swooper-maps",
+  "lane": "enforced",
+  "forbids": "foundation operation contracts importing domain config bags or using FoundationConfigSchema",
+  "why": "Foundation operation contracts must expose decomposed typed inputs rather than root config-bag surfaces.",
+  "remediate": null,
+  "message": "Remove foundation config-bag references from operation contract surfaces.",
+  "pathCoverage": [
+    {
+      "kind": "exact-path",
+      "patterns": [
+        "mods/mod-swooper-maps/src/domain/foundation/ops/*/contract.ts"
+      ]
+    }
+  ],
+  "scanRoots": [
+    "mods/mod-swooper-maps/src/domain/foundation/ops"
+  ],
+  "hookCheck": true,
+  "supportFiles": {
+    "baseline": ".habitat/civ7/mapgen/domains/foundation/rules/prohibit_foundation_op_contract_config_bags/baseline.json"
+  },
+  "runner": {
+    "name": "grit",
+    "files": {
+      "pattern": ".habitat/civ7/mapgen/domains/foundation/rules/prohibit_foundation_op_contract_config_bags/pattern.md"
+    },
+    "patternName": "prohibit_foundation_op_contract_config_bags"
+  }
+}
