--- .habitat/civ7/mapgen/domains/morphology/_remainder/prohibit_runtime_continent_step_tokens/rule.json
@@ -1,40 +1,0 @@
-{
-  "schemaVersion": 2,
-  "id": "prohibit_runtime_continent_step_tokens",
-  "title": "Prohibit Runtime Continent Step Tokens",
-  "placement": {
-    "niche": "civ7/mapgen/domains/morphology",
-    "blueprint": "_remainder",
-    "category": "execution"
-  },
-  "operation": {
-    "kind": "check"
-  },
-  "ownerProject": "mod-swooper-maps",
-  "lane": "enforced",
-  "forbids": "runtime continent identifiers and markLandmassId calls in morphology or hydrology implementation files",
-  "why": "Morphology and hydrology stages consume canonical artifacts instead of direct runtime continent marking.",
-  "remediate": null,
-  "message": "Route landmass identity through canonical artifacts instead of runtime continent step tokens.",
-  "pathCoverage": [
-    {
-      "kind": "exact-path",
-      "patterns": [
-        "mods/mod-swooper-maps/src/recipes/standard/stages/morphology*/**/*.ts",
-        "mods/mod-swooper-maps/src/recipes/standard/stages/hydrology-climate-baseline/**/*.ts"
-      ]
-    }
-  ],
-  "scanRoots": ["mods/mod-swooper-maps/src/recipes/standard/stages"],
-  "hookCheck": true,
-  "supportFiles": {
-    "baseline": ".habitat/civ7/mapgen/domains/morphology/_remainder/prohibit_runtime_continent_step_tokens/baseline.json"
-  },
-  "runner": {
-    "name": "grit",
-    "files": {
-      "pattern": ".habitat/civ7/mapgen/domains/morphology/_remainder/prohibit_runtime_continent_step_tokens/pattern.md"
-    },
-    "patternName": "prohibit_runtime_continent_step_tokens"
-  }
-}
