--- .habitat/civ7/mapgen/domains/morphology/_remainder/require_owned_domain_config_catalog_surfaces/rule.json
@@ -1,39 +1,0 @@
-{
-  "schemaVersion": 2,
-  "id": "require_owned_domain_config_catalog_surfaces",
-  "title": "Require Owned Domain Config Catalog Surfaces",
-  "placement": {
-    "niche": "civ7/mapgen/domains/morphology",
-    "blueprint": "_remainder",
-    "category": "contract"
-  },
-  "operation": {
-    "kind": "check"
-  },
-  "ownerProject": "mod-swooper-maps",
-  "lane": "enforced",
-  "forbids": "missing required morphology config facade exports or owner tag catalog tokens",
-  "why": "These residual positive validators assert exact current package surfaces until a future manifest-backed owner replaces them.",
-  "remediate": null,
-  "message": "Keep required morphology config exports and owner tag catalog tokens present.",
-  "pathCoverage": [
-    {
-      "kind": "exact-path",
-      "patterns": [
-        "mods/mod-swooper-maps/src/domain/morphology/config.ts",
-        "mods/mod-swooper-maps/src/recipes/standard/tags.ts"
-      ]
-    }
-  ],
-  "supportFiles": {
-    "baseline": ".habitat/civ7/mapgen/domains/morphology/_remainder/require_owned_domain_config_catalog_surfaces/baseline.json"
-  },
-  "runner": {
-    "name": "habitat",
-    "mode": "script",
-    "runtime": "node",
-    "files": {
-      "script": ".habitat/civ7/mapgen/domains/morphology/_remainder/require_owned_domain_config_catalog_surfaces/check.mjs"
-    }
-  }
-}
