+++ .habitat/civ7/mapgen/domains/morphology/rules/require_morphology_config_facade_exports/baseline.json
