+++ .habitat/civ7/mapgen/domains/narrative/rules/require_narrative_hotspot_overlay_owner/baseline.json
