+++ .habitat/civ7/mapgen/domains/narrative/rules/require_narrative_hotspot_overlay_owner/rule.json
@@ -1,0 +1,38 @@
+{
+  "schemaVersion": 2,
+  "id": "require_narrative_hotspot_overlay_owner",
+  "title": "Require Narrative Hotspot Overlay Owner",
+  "placement": {
+    "niche": "civ7/mapgen/domains/narrative",
+    "blueprint": "_self",
+    "category": "boundary"
+  },
+  "operation": {
+    "kind": "check"
+  },
+  "ownerProject": "mod-swooper-maps",
+  "lane": "enforced",
+  "forbids": "HOTSPOTS story overlay publishers outside the narrative hotspot owner",
+  "why": "HOTSPOTS overlay publication is narrative domain authority and must not be hidden inside morphology stage ownership.",
+  "remediate": null,
+  "message": "Keep HOTSPOTS story overlay publication owned by narrative tagging.",
+  "pathCoverage": [
+    {
+      "kind": "exact-path",
+      "patterns": [
+        "mods/mod-swooper-maps/src/domain/**/*.ts"
+      ]
+    }
+  ],
+  "supportFiles": {
+    "baseline": ".habitat/civ7/mapgen/domains/narrative/rules/require_narrative_hotspot_overlay_owner/baseline.json"
+  },
+  "runner": {
+    "name": "habitat",
+    "mode": "script",
+    "runtime": "node",
+    "files": {
+      "script": ".habitat/civ7/mapgen/domains/narrative/rules/require_narrative_hotspot_overlay_owner/check.mjs"
+    }
+  }
+}
