--- .habitat/civ7/mapgen/pipeline/swooper-maps-standard-recipe/_remainder/prohibit_legacy_plate_driver_and_plot_mountains_dependencies/rule.json
@@ -1,45 +1,0 @@
-{
-  "schemaVersion": 2,
-  "id": "prohibit_legacy_plate_driver_and_plot_mountains_dependencies",
-  "title": "Prohibit Legacy Plate Driver And Plot Mountains Dependencies",
-  "placement": {
-    "niche": "civ7/mapgen/pipeline/swooper-maps-standard-recipe",
-    "blueprint": "_remainder",
-    "category": "contract"
-  },
-  "operation": {
-    "kind": "check"
-  },
-  "ownerProject": "mod-swooper-maps",
-  "lane": "enforced",
-  "forbids": "legacy foundationPlates, foundationArtifacts.plates, tectonic-history, or tectonic-provenance dependencies in morphology contracts/source",
-  "why": "Morphology belt-driver contracts should use canonical belt driver artifacts rather than retired plate dependency surfaces.",
-  "remediate": null,
-  "message": "Use canonical belt-driver artifacts instead of legacy plate-driver dependencies.",
-  "pathCoverage": [
-    {
-      "kind": "exact-path",
-      "patterns": [
-        "mods/mod-swooper-maps/src/recipes/standard/stages/morphology-coasts/steps/landmassPlates.contract.ts",
-        "mods/mod-swooper-maps/src/recipes/standard/stages/morphology-features/steps/mountains.contract.ts",
-        "mods/mod-swooper-maps/src/recipes/standard/stages/map-morphology/steps/plotMountains.ts"
-      ]
-    }
-  ],
-  "scanRoots": [
-    "mods/mod-swooper-maps/src/recipes/standard/stages/morphology-coasts/steps/landmassPlates.contract.ts",
-    "mods/mod-swooper-maps/src/recipes/standard/stages/morphology-features/steps/mountains.contract.ts",
-    "mods/mod-swooper-maps/src/recipes/standard/stages/map-morphology/steps/plotMountains.ts"
-  ],
-  "hookCheck": true,
-  "supportFiles": {
-    "baseline": ".habitat/civ7/mapgen/pipeline/swooper-maps-standard-recipe/_remainder/prohibit_legacy_plate_driver_and_plot_mountains_dependencies/baseline.json"
-  },
-  "runner": {
-    "name": "grit",
-    "files": {
-      "pattern": ".habitat/civ7/mapgen/pipeline/swooper-maps-standard-recipe/_remainder/prohibit_legacy_plate_driver_and_plot_mountains_dependencies/pattern.md"
-    },
-    "patternName": "prohibit_legacy_plate_driver_and_plot_mountains_dependencies"
-  }
-}
