+++ .habitat/civ7/mapgen/pipeline/swooper-maps-standard-recipe/stages/foundation/rules/prohibit_foundation_stage_sentinel_passthrough/pattern.md
@@ -11,7 +11,7 @@ language js(typescript)
 or {
   contains "FOUNDATION_STUDIO_STEP_CONFIG_IDS",
   contains "FOUNDATION_STEP_IDS",
-  contains "advancedRecord[stepId]",
+  contains r"\badvancedRecord\s*\[\s*stepId\s*\]",
   contains "__studioUiMetaSentinelPath"
 } where {
   $filename <: r".*mods/mod-swooper-maps/src/recipes/standard/stages/foundation-(?:lithosphere|mantle|orogeny|projection|tectonics)/index\.ts$"
