+++ .habitat/civ7/mapgen/pipeline/swooper-maps-standard-recipe/stages/foundation/rules/prohibit_foundation_step_contract_config_bags/pattern.md
@@ -1,0 +1,19 @@
+---
+level: error
+---
+# Prohibit Foundation Step Contract Config Bags
+
+Foundation recipe step contracts must not depend on root config-bag imports or schemas.
+
+```grit
+language js(typescript)
+
+or {
+  import_statement(source=$source) where {
+    $source <: r".*@mapgen/domain/config.*"
+  },
+  `FoundationConfigSchema`
+} where {
+  $filename <: r".*mods/mod-swooper-maps/src/recipes/standard/stages/foundation(?:-[^/]+)?/.*contract\.ts$"
+}
+```
