+++ .habitat/civ7/mapgen/pipeline/swooper-maps-standard-recipe/stages/map/rules/prohibit_map_morphology_legacy_plate_driver_dependencies/baseline.json
@@ -1,0 +1,1 @@
+[]
