+++ .habitat/civ7/mapgen/pipeline/swooper-maps-standard-recipe/stages/map/rules/prohibit_map_morphology_legacy_plate_driver_dependencies/pattern.md
@@ -1,0 +1,19 @@
+---
+level: error
+---
+# Prohibit Map Morphology Legacy Plate Driver Dependencies
+
+Map-morphology mountain plotting source must not depend on retired plate-driver
+surfaces.
+
+```grit
+language js(typescript)
+
+or {
+  contains r"\bfoundationPlates\b",
+  contains r"\bfoundationArtifacts\.plates\b"
+} where {
+  $filename <: r".*mods/mod-swooper-maps/src/recipes/standard/stages/map-morphology/steps/plotMountains\.ts$"
+}
+```
+
