+++ .habitat/civ7/mapgen/pipeline/swooper-maps-standard-recipe/stages/map/rules/prohibit_map_morphology_legacy_plate_driver_dependencies/rule.json
@@ -1,0 +1,41 @@
+{
+  "schemaVersion": 2,
+  "id": "prohibit_map_morphology_legacy_plate_driver_dependencies",
+  "title": "Prohibit Map Morphology Legacy Plate Driver Dependencies",
+  "placement": {
+    "niche": "civ7/mapgen/pipeline/swooper-maps-standard-recipe/stages/map",
+    "blueprint": "_self",
+    "category": "contract"
+  },
+  "operation": {
+    "kind": "check"
+  },
+  "ownerProject": "mod-swooper-maps",
+  "lane": "enforced",
+  "forbids": "legacy foundation plate-driver dependencies in map-morphology mountain plotting source",
+  "why": "Map-morphology projection source should consume canonical morphology outputs rather than retired foundation plate surfaces.",
+  "remediate": null,
+  "message": "Use canonical morphology/projected inputs instead of legacy plate-driver dependencies.",
+  "pathCoverage": [
+    {
+      "kind": "exact-path",
+      "patterns": [
+        "mods/mod-swooper-maps/src/recipes/standard/stages/map-morphology/steps/plotMountains.ts"
+      ]
+    }
+  ],
+  "scanRoots": [
+    "mods/mod-swooper-maps/src/recipes/standard/stages/map-morphology/steps/plotMountains.ts"
+  ],
+  "hookCheck": true,
+  "supportFiles": {
+    "baseline": ".habitat/civ7/mapgen/pipeline/swooper-maps-standard-recipe/stages/map/rules/prohibit_map_morphology_legacy_plate_driver_dependencies/baseline.json"
+  },
+  "runner": {
+    "name": "grit",
+    "files": {
+      "pattern": ".habitat/civ7/mapgen/pipeline/swooper-maps-standard-recipe/stages/map/rules/prohibit_map_morphology_legacy_plate_driver_dependencies/pattern.md"
+    },
+    "patternName": "prohibit_map_morphology_legacy_plate_driver_dependencies"
+  }
+}
