+++ .habitat/civ7/mapgen/pipeline/swooper-maps-standard-recipe/stages/morphology/rules/prohibit_morphology_contract_legacy_plate_driver_dependencies/baseline.json
@@ -1,0 +1,1 @@
+[]
