+++ .habitat/civ7/mapgen/pipeline/swooper-maps-standard-recipe/stages/morphology/rules/prohibit_morphology_contract_legacy_plate_driver_dependencies/pattern.md
@@ -1,0 +1,20 @@
+---
+level: error
+---
+# Prohibit Morphology Contract Legacy Plate Driver Dependencies
+
+Morphology contracts must not depend on retired plate-driver artifacts.
+
+```grit
+language js(typescript)
+
+or {
+  contains r"mapArtifacts\.foundationPlates" where {
+    $filename <: r".*mods/mod-swooper-maps/src/recipes/standard/stages/morphology-coasts/steps/landmassPlates\.contract\.ts$"
+  },
+  contains r"mapArtifacts\.foundation(?:TectonicHistoryTiles|TectonicProvenanceTiles|Plates)" where {
+    $filename <: r".*mods/mod-swooper-maps/src/recipes/standard/stages/morphology-features/steps/mountains\.contract\.ts$"
+  }
+}
+```
+
