+++ .habitat/civ7/mapgen/pipeline/swooper-maps-standard-recipe/stages/morphology/rules/prohibit_morphology_runtime_continent_step_tokens/baseline.json
@@ -1,0 +1,1 @@
+[]
