+++ .habitat/civ7/mapgen/pipeline/swooper-maps-standard-recipe/stages/morphology/rules/prohibit_morphology_stage_legacy_effect_gates/baseline.json
@@ -1,0 +1,1 @@
+[]
