+++ .habitat/civ7/mapgen/pipeline/swooper-maps-standard-recipe/stages/morphology/rules/prohibit_morphology_stage_legacy_effect_gates/pattern.md
@@ -1,0 +1,15 @@
+---
+level: error
+---
+# Prohibit Morphology Stage Legacy Effect Gates
+
+Morphology stages must not use retired engine landmass or coastlines effect gates.
+
+```grit
+language js(typescript)
+
+contains r"(?:landmassApplied|coastlinesApplied|effect:engine\.landmassApplied|effect:engine\.coastlinesApplied)" where {
+  $filename <: r".*mods/mod-swooper-maps/src/recipes/standard/stages/(?:morphology-coasts|morphology-routing|morphology-erosion|morphology-features)/.*\.ts$"
+}
+```
+
